+++ changelog.md
@@ -4,3 +4,5 @@ a release
 -->
 
 ## Notable Changes
+
+- **Added** a new error which describes that a gpg key could not be loaded